Move CTransfer_MixingPlaneInterface
