+++ SU2_CFD/include/transfer/CTransfer.hpp
@@ -573,84 +573,3 @@ public:
 
 
 };
-
-/*!
- * \class CTransfer_MixingPlaneInterface
- * \brief Transfer average variables needed for MixingPlane computation from a generic zone into another one
- * \author S. Vitale
- * \version 6.2.0 "Falcon"
- */
-
-
-class CTransfer_MixingPlaneInterface : public CTransfer {
-
-public:
-
-	/*!
-	 * \brief Constructor of the class.
-	 */
-	CTransfer_MixingPlaneInterface(void);
-
-	/*!
-	 * \overload
-	 * \param[in] val_nVar - Number of variables that need to be transferred.
-	 * \param[in] config - Definition of the particular problem.
-	 */
-	CTransfer_MixingPlaneInterface(unsigned short val_nVar, unsigned short val_nConst, CConfig *donor_config, CConfig *target_config);
-
-	/*!
-	 * \brief Destructor of the class.
-	 */
-	virtual ~CTransfer_MixingPlaneInterface(void);
-
-
-/*!
- * \brief Initialize quantities for spanwise sections for interpolation.
- * \param[in] donor_config - Definition of the problem at the donor mesh.
- * \param[in] target_config - Definition of the problem at the target mesh.
- */
-void SetSpanWiseLevels(CConfig *donor_config, CConfig *target_config);
-
-	/*!
-	 * \brief Retrieve the variable that will be sent from donor mesh to target mesh.
-	 * \param[in] donor_solution - Solution from the donor mesh.
-	 * \param[in] donor_geometry - Geometry of the donor mesh.
-	 * \param[in] donor_config - Definition of the problem at the donor mesh.
-	 * \param[in] Marker_Donor - Index of the donor marker.
-	 * \param[in] Vertex_Donor - Index of the donor vertex.
-	 * \param[in] Point_Donor - Index of the donor point.
-	 */
-	void GetDonor_Variable(CSolver *donor_solution, CGeometry *donor_geometry, CConfig *donor_config,
-						   unsigned long Marker_Donor, unsigned long val_Span, unsigned long Point_Donor);
-
-	/*!
-	 * \brief Set the variable that has been received from the target mesh into the target mesh.
-	 * \param[in] target_solution - Solution from the target mesh.
-	 * \param[in] target_geometry - Geometry of the target mesh.
-	 * \param[in] target_config - Definition of the problem at the target mesh.
-	 * \param[in] Marker_Target - Index of the target marker.
-	 * \param[in] Vertex_Target - Index of the target vertex.
-	 * \param[in] Point_Target - Index of the target point.
-	 */
-	void SetTarget_Variable(CSolver *target_solution, CGeometry *target_geometry, CConfig *target_config,
-							unsigned long Marker_Target, unsigned long val_Span, unsigned long Point_Target);
-
-	/*!
-	 * \brief Store all the turboperformance in the solver in ZONE_0.
-	 * \param[in] donor_solution  - Solution from the donor mesh.
-	 * \param[in] target_solution - Solution from the target mesh.
-	 * \param[in] donorZone       - counter of the donor solution
-	 */
-	void SetAverageValues(CSolver *donor_solution, CSolver *target_solution, unsigned short donorZone);
-
-	/*!
-	 * \brief Store all the turboperformance in the solver in ZONE_0.
-	 * \param[in] donor_geometry  - Solution from the donor mesh.
-	 * \param[in] target_geometry - Solution from the target mesh.
-	 * \param[in] donorZone       - counter of the donor solution
-	 */
-	void SetAverageTurboGeoValues(CGeometry *donor_geometry, CGeometry *target_geometry, unsigned short donorZone);
-
-
-
-};
